refactor: 공통 패키지 구조에 맞춰 변경
