+++ src/test/java/com/careerpirates/resumate/folder/infrastructure/FolderRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.careerpirates.resumate.folder.infrasturcture;
+package com.careerpirates.resumate.folder.infrastructure;
 
 import com.careerpirates.resumate.folder.domain.Folder;
 import org.junit.jupiter.api.AfterEach;
@@ -7,13 +7,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
 import java.util.List;
 
 import static com.careerpirates.resumate.folder.factory.FolderTestFactory.createDefaultFolders;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 
 @SpringBootTest
